refactor some components as models for the rest one
