+++ src/UI/extension/ExtensionLoginButton/ExtensionLoginButton.tsx
@@ -1,98 +1,16 @@
-import React, { ReactNode } from 'react';
-import classNames from 'classnames';
-import {
-  CHROME_EXTENSION_LINK,
-  DataTestIdsEnum,
-  FIREFOX_ADDON_LINK
-} from 'constants/index';
-import { withStyles, WithStylesImportType } from 'hocs/withStyles';
-import { useExtensionLogin } from 'hooks/login/useExtensionLogin';
-import { getIsNativeAuthSingingForbidden } from 'services/nativeAuth/helpers';
-import { LoginButton } from 'UI/LoginButton/LoginButton';
-import { OnProviderLoginType } from '../../../types';
-import { WithClassnameType } from '../../types';
-import { getIsExtensionAvailable } from '../helpers';
+import React from 'react';
+import globalStyles from 'assets/sass/main.scss';
+import styles from './extensionLoginButton.styles.scss';
+import { ExtensionLoginButtonComponent } from './ExtensionLoginButtonComponent';
 
-export interface ExtensionLoginButtonPropsType
-  extends WithClassnameType,
-    OnProviderLoginType {
-  children?: ReactNode;
-  buttonClassName?: string;
-  loginButtonText?: string;
-  disabled?: boolean;
-}
-
-const isExtensionAvailable = getIsExtensionAvailable();
-
-const ExtensionLoginButtonComponent: (
-  props: ExtensionLoginButtonPropsType & WithStylesImportType
-) => JSX.Element = ({
-  token,
-  className = 'dapp-extension-login',
-  children,
-  callbackRoute,
-  buttonClassName = 'dapp-default-login-button',
-  nativeAuth,
-  loginButtonText = 'MultiversX DeFi Wallet',
-  onLoginRedirect,
-  disabled,
-  'data-testid': dataTestId = DataTestIdsEnum.extensionLoginButton,
-  globalStyles,
-  styles
-}) => {
-  const [onInitiateLogin] = useExtensionLogin({
-    callbackRoute,
-    token,
-    onLoginRedirect,
-    nativeAuth
-  });
-  const disabledConnectButton = getIsNativeAuthSingingForbidden(token);
-  const isFirefox = navigator.userAgent.indexOf('Firefox') != -1;
-  const classes = {
-    wrapper: classNames(
-      globalStyles?.btn,
-      globalStyles?.btnPrimary,
-      globalStyles?.px4,
-      globalStyles?.m1,
-      globalStyles?.mx3,
-      styles?.noExtensionButtonWrapper,
-      {
-        [buttonClassName]: buttonClassName != null
-      },
-      className
-    ),
-    loginText: classNames(styles?.loginText, styles?.noExtensionButtonContent),
-    wrapperClassName: className
-  };
-
-  const handleLogin = () => {
-    onInitiateLogin();
-  };
-
-  return !isExtensionAvailable ? (
-    <a
-      rel='noreferrer'
-      href={isFirefox ? FIREFOX_ADDON_LINK : CHROME_EXTENSION_LINK}
-      target='_blank'
-      className={classes.wrapper}
-    >
-      {children || <span className={classes.loginText}>{loginButtonText}</span>}
-    </a>
-  ) : (
-    <LoginButton
-      onLogin={handleLogin}
-      className={className}
-      btnClassName={buttonClassName}
-      text={loginButtonText}
-      disabled={disabled || disabledConnectButton}
-      data-testid={dataTestId}
-    >
-      {children}
-    </LoginButton>
+export const ExtensionLoginButton: typeof ExtensionLoginButtonComponent = (
+  props
+) => {
+  return (
+    <ExtensionLoginButtonComponent
+      {...props}
+      globalStyles={globalStyles}
+      styles={styles}
+    />
   );
 };
-
-export const ExtensionLoginButton = withStyles(ExtensionLoginButtonComponent, {
-  local: () =>
-    import('UI/extension/ExtensionLoginButton/extensionLoginButton.styles.scss')
-});
