+++ src/UI/extension/ExtensionLoginButton/index.ts
@@ -1,1 +1,2 @@
 export * from './ExtensionLoginButton';
+export * from './types';
