+++ src/UI/extension/ExtensionLoginButton/types.ts
@@ -1,0 +1,12 @@
+import { ReactNode } from 'react';
+import { OnProviderLoginType } from '../../../types';
+import { WithClassnameType } from '../../types';
+
+export interface ExtensionLoginButtonPropsType
+  extends WithClassnameType,
+    OnProviderLoginType {
+  children?: ReactNode;
+  buttonClassName?: string;
+  loginButtonText?: string;
+  disabled?: boolean;
+}
